+++ packages/ensnode-schema/src/ponder.schema.ts
@@ -4,3 +4,4 @@
 export * from "./subgraph.schema";
 export * from "./resolver-records.schema";
 export * from "./seaport.schema";
+export * from "./resolver-relations.schema";